adicionando função para alterar

+++ Painter/app/src/main/java/com/prado/painter/persistence/ServicesDAO.java
@@ -22,7 +22,7 @@ public interface ServicesDAO {
     void update(Service service);
 
     @Query("SELECT * FROM service WHERE id = :id")
-    Service queryForId(long id);
+    Service queryForId(int id);
 
     @Query("SELECT * FROM service ")
     List<Service> queryAll();
